improve handling of CD-ROM drive readiness
Implement our own code to close the tray if necessary and wait for the
device to become ready.  The kernel's built-in handling of this on
open() will close the tray, but fails if the drive doesn't become
immediately ready afterwards.

Also add an --eject flag to archive.py to cause it to eject the disc
after archiving.