fix bugs handling unicode file names
Fix crashes when handling non-ASCII characters in directory names and in
track titles.  In two places filenames were inadvertantly left as, or
converted to, a plain non-unicode byte string.  This caused decode
errors when we tried to concatenate them with Unicode strings, and
python attempted to automatically encode them as ASCII.

In album_dir.py, the input path was left as a plain byte string, rather
than being decoded as UTF-8.

In rename.py the safe_filename() function was inadvertantly converting
unicode input strings to plain byte strings in output, due to passing
the input through a non-unicode format string.